refactored code; updated dir structure
