--- include/ast/nodes/identifier.h
@@ -1,8 +1,0 @@
-#ifndef AST_NODES_IDENTIFIER_H
-#define AST_NODES_IDENTIFIER_H
-
-#include "ast.h"
-
-void AST_Identifier_free(AST_Identifier_t *identifier);
-
-#endif
