--- include/ast/nodes/statements.h
@@ -1,8 +1,0 @@
-#ifndef AST_NODES_STATEMENTS_H
-#define AST_NODES_STATEMENTS_H
-
-#include "ast.h"
-
-void AST_Statements_free(AST_Statements_t *statements);
-
-#endif
