+++ src/ast.c
@@ -1,8 +1,5 @@
 #include <stdlib.h>
 
 #include "ast.h"
-
 #include "ast/nodes.c.h"
-#include "ast/nodes/identifier.c.h"
-#include "ast/nodes/statements.c.h"
 #include "ast/util.c.h"
