+++ src/ast/nodes/destroy.c.h
@@ -1,10 +1,8 @@
-#ifndef AST_NODES_STATEMENTS_C_H
-#define AST_NODES_STATEMENTS_C_H
+#ifndef AST_NODES_DESTROY_C_H
+#define AST_NODES_DESTROY_C_H
 
 #include <stdlib.h>
-
 #include "ast/nodes.h"
-#include "ast/nodes/statements.h"
 
 void AST_Statements_free(AST_Statements_t *statements)
 {
@@ -18,6 +16,12 @@ void AST_Statements_free(AST_Statements_t *statements)
     }
 }
 
+void AST_Identifier_free(AST_Identifier_t *identifier)
+{
+    free(identifier->identifier_name);
+    free(identifier);
+}
+
 #else
-    #warning re-inclusion of module 'ast/nodes/statements.c.h'
+    #warning re-inclusion of module 'ast/nodes/destroy.c.h'
 #endif
